+++ lib/segment/segment.dart
@@ -3,7 +3,6 @@ library segment;
 
 import 'package:uuid/uuid.dart' show Uuid;
 import '../context/context.dart' show Context;
-import '../event/event.dart' show EventBuffer;
 import '../lifecycle/lifecycle.dart' show AppLifecycle;
 import '../parser/parser.dart' show AnalyticsParser;
 import '../store/store.dart' show Store;
@@ -18,16 +17,13 @@ abstract class Segment {
   /// @nodoc
   Segment(dynamic properties)
       : _appLifecycle = AppLifecycle().toString(),
-        _buffer = EventBuffer(),
+        _anonymousId = Store().anonymousId,
+        _context = Context().toMap(),
         _properties = AnalyticsParser(properties).toJson(),
+        _sessionId = Store().sessionId,
         _timestamp = DateTime.now().toUtc().toIso8601String(),
-        _tzOffsetHours = DateTime.now().timeZoneOffset.inHours {
-    _anonymousId = _buffer.defer<String>(() => Store().anonymousId);
-    _context = _buffer.defer<Map<String, dynamic>>(() => Context().toMap());
-    _orgId = _buffer.defer<String>(() => Store().orgId);
-    _sessionId = _buffer.defer<String>(() => Store().sessionId);
-    _userId = _buffer.defer<String>(() => Store().userId);
-  }
+        _tzOffsetHours = DateTime.now().timeZoneOffset.inHours,
+        _userId = Store().userId;
 
   static final _dartEnv = dartEnv();
 
@@ -35,21 +31,25 @@ abstract class Segment {
   static String _nextMessageId;
 
   final String _appLifecycle;
-  final EventBuffer _buffer;
+  final Future<String> _anonymousId;
+  final Future<Map<String, dynamic>> _context;
   final Map<String, dynamic> _properties;
+  final Future<String> _sessionId;
   final String _timestamp;
   final int _tzOffsetHours;
-
-  Future<String> _anonymousId;
-  Future<Map<String, dynamic>> _context;
-  Future<String> _orgId;
-  Future<String> _sessionId;
-  Future<String> _userId;
+  final Future<String> _userId;
 
   /// @nodoc
   Future<Map<String, dynamic>> toMap() async {
+    await Future<void>.delayed(Duration(microseconds: 1));
+
     final messageId = _nextMessageId ?? Uuid().v4();
+    final previousMessageId = _previousMessageId ?? messageId;
+
+    _previousMessageId = messageId;
     _nextMessageId = Uuid().v4();
+
+    final nextMessageId = _nextMessageId;
 
     final payload = <String, dynamic>{
       'anonymousId': await _anonymousId,
@@ -57,12 +57,12 @@ abstract class Segment {
       'messageId': messageId,
       'properties': <String, dynamic>{
         ..._properties,
-        'orgId': await _orgId,
+        'orgId': await Store().orgId,
         'sdk': <String, dynamic>{
           'appLifecycle': _appLifecycle,
           'dartEnv': _dartEnv,
-          'nextMessageId': _nextMessageId,
-          'previousMessageId': _previousMessageId ?? messageId,
+          'nextMessageId': nextMessageId,
+          'previousMessageId': previousMessageId,
           'sessionId': await _sessionId,
           'tzOffsetHours': _tzOffsetHours
         }
@@ -70,10 +70,6 @@ abstract class Segment {
       'timestamp': _timestamp,
       'userId': await _userId
     };
-
-    _previousMessageId = messageId;
-
-    await _buffer.destroy();
 
     return payload;
   }