+++ lib/version_control.dart
@@ -5,7 +5,7 @@ library version_control;
 const String sdkName = 'flutter_analytics';
 
 /// @nodoc
-const String sdkVersion = '7.2.2';
+const String sdkVersion = '7.2.3';
 
 /// @nodoc
 const sdkPackage = <String, dynamic>{'name': sdkName, 'version': sdkVersion};