+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [v7.2.3](https://github.com/oakromulo/flutter_analytics/tree/v7.2.3) - 2020-02-27
+
+- Fix possible `messageId` inconsistency.
+- Update `location` dependency.
+
 ## [v7.2.2](https://github.com/oakromulo/flutter_analytics/tree/v7.2.2) - 2020-02-25
 
 - Enforce non-null organization header.